split cli spike into separate files
