+++ src/cli/gameOver.js
@@ -1,0 +1,9 @@
+const display = require('./display')
+
+module.exports = (options, io) => {
+  const { state, messages } = options
+  const { write } = io
+
+  write(display.board(state))
+  write(messages.ending)
+}
